Creating a branch for a better support of value oracles.
The idea is that the matching phase does not have to establish images for value nodes whose value is provided through an oracle, provided that there are no constraints on the values. Instead these values can be filled in at rule application (and stored in the target graph rather than the event).